+++ sageopt/__init__.py
@@ -17,18 +17,18 @@ from sageopt import coniclifts
 from sageopt import relaxations
 from sageopt import symbolic
 
+from sageopt.relaxations import local_refine, local_refine_polys_from_sigs
+from sageopt.relaxations import conditional_sage_data
+from sageopt.relaxations import sage_feasibility, sage_multiplier_search
+
 from sageopt.symbolic.signomials import Signomial, standard_sig_monomials
-from sageopt.relaxations.sage_sigs import conditional_sage_data
-from sageopt.relaxations.sage_sigs import sage_primal, sage_dual, sage_feasibility, sage_multiplier_search
-from sageopt.relaxations.sage_sigs import constrained_sage_primal, constrained_sage_dual
-from sageopt.relaxations.sig_solution_recovery import dual_solution_recovery, local_refinement
+from sageopt.relaxations import sig_primal, sig_dual
+from sageopt.relaxations import sig_constrained_primal, sig_constrained_dual
+from sageopt.relaxations import sig_solrec
 
 from sageopt.symbolic.polynomials import Polynomial, standard_poly_monomials
-from sageopt.relaxations.sage_polys import conditional_sage_poly_data
-from sageopt.relaxations.sage_polys import sage_poly_primal, sage_poly_dual
-from sageopt.relaxations.sage_polys import sage_poly_feasibility, sage_poly_multiplier_search
-from sageopt.relaxations.sage_polys import constrained_sage_poly_primal, constrained_sage_poly_dual
-from sageopt.relaxations.poly_solution_recovery import dual_solution_recovery as dual_poly_solution_recovery
-from sageopt.relaxations.poly_solution_recovery import local_refinement as local_poly_refinement
+from sageopt.relaxations import poly_primal, poly_dual
+from sageopt.relaxations import poly_constrained_primal, poly_constrained_dual
+from sageopt.relaxations import poly_solrec
 
 __version__ = '0.3.0'